Fix duplicated classes
