+++ lib/models/customer.dart
@@ -31,6 +31,10 @@
 
  */
 
+import 'billing.dart';
+import 'shipping.dart';
+import 'meta_data.dart';
+
 class WooCustomer {
   int id;
   String dateCreated;
@@ -47,7 +51,7 @@ class WooCustomer {
   Shipping shipping;
   bool isPayingCustomer;
   String avatarUrl;
-  List<WooCustomerMetaData> metaData;
+  List<MetaData> metaData;
   Links links;
 
   WooCustomer(
@@ -88,7 +92,7 @@ class WooCustomer {
     isPayingCustomer = json['is_paying_customer'];
     avatarUrl = json['avatar_url'];
     metaData =
-        (json['meta_data'] as List).map((i) => WooCustomerMetaData.fromJson(i)).toList();
+        (json['meta_data'] as List).map((i) => MetaData.fromJson(i)).toList();
     links = json['_links'] != null ? new Links.fromJson(json['_links']) : null;
   }
 
@@ -127,127 +131,6 @@ class WooCustomer {
     return data;
   }
   @override toString() => this.toJson().toString();
-}
-
-class WooCustomerMetaData {
-  final int id;
-  final String key;
-  final dynamic value;
-
-  WooCustomerMetaData(this.id, this.key, this.value);
-
-  WooCustomerMetaData.fromJson(Map<String, dynamic> json)
-      : id = json['id'],
-        key = json['key'],
-        value = json['value'];
-
-  Map<String, dynamic> toJson() => {'id': id, 'key': key, 'value': value};
-}
-
-class Billing {
-  String firstName;
-  String lastName;
-  String company;
-  String address1;
-  String address2;
-  String city;
-  String state;
-  String postcode;
-  String country;
-  String email;
-  String phone;
-
-  Billing(
-      {this.firstName,
-      this.lastName,
-      this.company,
-      this.address1,
-      this.address2,
-      this.city,
-      this.state,
-      this.postcode,
-      this.country,
-      this.email,
-      this.phone});
-
-  Billing.fromJson(Map<String, dynamic> json) {
-    firstName = json['first_name'];
-    lastName = json['last_name'];
-    company = json['company'];
-    address1 = json['address_1'];
-    address2 = json['address_2'];
-    city = json['city'];
-    state = json['state'];
-    postcode = json['postcode'];
-    country = json['country'];
-    email = json['email'];
-    phone = json['phone'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['first_name'] = this.firstName;
-    data['last_name'] = this.lastName;
-    data['company'] = this.company;
-    data['address_1'] = this.address1;
-    data['address_2'] = this.address2;
-    data['city'] = this.city;
-    data['state'] = this.state;
-    data['postcode'] = this.postcode;
-    data['country'] = this.country;
-    data['email'] = this.email;
-    data['phone'] = this.phone;
-    return data;
-  }
-}
-
-class Shipping {
-  String firstName;
-  String lastName;
-  String company;
-  String address1;
-  String address2;
-  String city;
-  String state;
-  String postcode;
-  String country;
-
-  Shipping(
-      {this.firstName,
-      this.lastName,
-      this.company,
-      this.address1,
-      this.address2,
-      this.city,
-      this.state,
-      this.postcode,
-      this.country});
-
-  Shipping.fromJson(Map<String, dynamic> json) {
-    firstName = json['first_name'];
-    lastName = json['last_name'];
-    company = json['company'];
-    address1 = json['address_1'];
-    address2 = json['address_2'];
-    city = json['city'];
-    state = json['state'];
-    postcode = json['postcode'];
-    country = json['country'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['first_name'] = this.firstName;
-    data['last_name'] = this.lastName;
-    data['company'] = this.company;
-    data['address_1'] = this.address1;
-    data['address_2'] = this.address2;
-    data['city'] = this.city;
-    data['state'] = this.state;
-    data['postcode'] = this.postcode;
-    data['country'] = this.country;
-    return data;
-  }
 }
 
 class Links {
