+++ lib/models/meta_data.dart
@@ -1,0 +1,21 @@
+class MetaData {
+  int id;
+  String key;
+  String value;
+
+  MetaData({this.id, this.key, this.value});
+
+  MetaData.fromJson(Map<String, dynamic> json) {
+    id = json['id'];
+    key = json['key'];
+    value = json['value'].toString();
+  }
+
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = new Map<String, dynamic>();
+    data['id'] = this.id;
+    data['key'] = this.key;
+    data['value'] = this.value;
+    return data;
+  }
+}
