+++ lib/models/order.dart
@@ -31,7 +31,11 @@
 
  */
 
-import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart';
+import 'meta_data.dart';
+import 'billing.dart';
+import 'shipping.dart';
+
 class WooOrder {
   int id;
   int parentId;
@@ -77,7 +81,8 @@ class WooOrder {
   Links links;
 
   WooOrder(
-      {@required this.id,
+      {@required
+      this.id,
       this.parentId,
       this.number,
       this.orderKey,
@@ -378,134 +383,6 @@ class FeeLineTax {
     data['tax_total'] = this.taxTotal;
     data['shipping_tax_total'] = this.shippingTaxTotal;
     data['meta_data'] = this.metaData;
-    return data;
-  }
-}
-
-class Billing {
-  String firstName;
-  String lastName;
-  String company;
-  String address1;
-  String address2;
-  String city;
-  String state;
-  String postcode;
-  String country;
-  String email;
-  String phone;
-
-  Billing(
-      {this.firstName,
-      this.lastName,
-      this.company,
-      this.address1,
-      this.address2,
-      this.city,
-      this.state,
-      this.postcode,
-      this.country,
-      this.email,
-      this.phone});
-
-  Billing.fromJson(Map<String, dynamic> json) {
-    firstName = json['first_name'];
-    lastName = json['last_name'];
-    company = json['company'];
-    address1 = json['address_1'];
-    address2 = json['address_2'];
-    city = json['city'];
-    state = json['state'];
-    postcode = json['postcode'];
-    country = json['country'];
-    email = json['email'];
-    phone = json['phone'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['first_name'] = this.firstName;
-    data['last_name'] = this.lastName;
-    data['company'] = this.company;
-    data['address_1'] = this.address1;
-    data['address_2'] = this.address2;
-    data['city'] = this.city;
-    data['state'] = this.state;
-    data['postcode'] = this.postcode;
-    data['country'] = this.country;
-    data['email'] = this.email;
-    data['phone'] = this.phone;
-    return data;
-  }
-}
-
-class Shipping {
-  String firstName;
-  String lastName;
-  String company;
-  String address1;
-  String address2;
-  String city;
-  String state;
-  String postcode;
-  String country;
-
-  Shipping(
-      {this.firstName,
-      this.lastName,
-      this.company,
-      this.address1,
-      this.address2,
-      this.city,
-      this.state,
-      this.postcode,
-      this.country});
-
-  Shipping.fromJson(Map<String, dynamic> json) {
-    firstName = json['first_name'];
-    lastName = json['last_name'];
-    company = json['company'];
-    address1 = json['address_1'];
-    address2 = json['address_2'];
-    city = json['city'];
-    state = json['state'];
-    postcode = json['postcode'];
-    country = json['country'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['first_name'] = this.firstName;
-    data['last_name'] = this.lastName;
-    data['company'] = this.company;
-    data['address_1'] = this.address1;
-    data['address_2'] = this.address2;
-    data['city'] = this.city;
-    data['state'] = this.state;
-    data['postcode'] = this.postcode;
-    data['country'] = this.country;
-    return data;
-  }
-}
-
-class MetaData {
-  int id;
-  String key;
-  String value;
-
-  MetaData({this.id, this.key, this.value});
-
-  MetaData.fromJson(Map<String, dynamic> json) {
-    id = json['id'];
-    key = json['key'];
-    value = json['value'].toString();
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['id'] = this.id;
-    data['key'] = this.key;
-    data['value'] = this.value;
     return data;
   }
 }
