+++ lib/models/shipping.dart
@@ -1,0 +1,48 @@
+class Shipping {
+  String firstName;
+  String lastName;
+  String company;
+  String address1;
+  String address2;
+  String city;
+  String state;
+  String postcode;
+  String country;
+
+  Shipping(
+      {this.firstName,
+        this.lastName,
+        this.company,
+        this.address1,
+        this.address2,
+        this.city,
+        this.state,
+        this.postcode,
+        this.country});
+
+  Shipping.fromJson(Map<String, dynamic> json) {
+    firstName = json['first_name'];
+    lastName = json['last_name'];
+    company = json['company'];
+    address1 = json['address_1'];
+    address2 = json['address_2'];
+    city = json['city'];
+    state = json['state'];
+    postcode = json['postcode'];
+    country = json['country'];
+  }
+
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = new Map<String, dynamic>();
+    data['first_name'] = this.firstName;
+    data['last_name'] = this.lastName;
+    data['company'] = this.company;
+    data['address_1'] = this.address1;
+    data['address_2'] = this.address2;
+    data['city'] = this.city;
+    data['state'] = this.state;
+    data['postcode'] = this.postcode;
+    data['country'] = this.country;
+    return data;
+  }
+}
